+++ src/QueryCat.Cli/Commands/Options/QueryOptionsBinder.cs
@@ -15,6 +15,7 @@ internal class QueryOptionsBinder : BinderBase<QueryOptions>
     private readonly Option<int> _analyzeRowsOption;
     private readonly Option<string?> _columnsSeparatorOption;
     private readonly Option<bool> _disableCacheOption;
+    private readonly Option<bool> _noHeaderOption;
 
     public QueryOptionsBinder(
         Option<int> maxErrorsOption,
@@ -24,7 +25,8 @@ internal class QueryOptionsBinder : BinderBase<QueryOptions>
         Option<int> pageSizeOption, Option<TextTableOutput.Style> outputStyleOption,
         Option<int> analyzeRowsOption,
         Option<string?> columnsSeparatorOption,
-        Option<bool> disableCacheOption)
+        Option<bool> disableCacheOption,
+        Option<bool> noHeaderOption)
     {
         _maxErrorsOption = maxErrorsOption;
         _statisticOption = statisticOption;
@@ -42,6 +44,7 @@ internal class QueryOptionsBinder : BinderBase<QueryOptions>
         _analyzeRowsOption = analyzeRowsOption;
         _columnsSeparatorOption = columnsSeparatorOption;
         _disableCacheOption = disableCacheOption;
+        _noHeaderOption = noHeaderOption;
     }
 
     /// <inheritdoc />
@@ -58,6 +61,7 @@ internal class QueryOptionsBinder : BinderBase<QueryOptions>
             AnalyzeRows = bindingContext.ParseResult.GetValueForOption(_analyzeRowsOption),
             ColumnsSeparator = bindingContext.ParseResult.GetValueForOption(_columnsSeparatorOption),
             DisableCache = bindingContext.ParseResult.GetValueForOption(_disableCacheOption),
+            NoHeader = bindingContext.ParseResult.GetValueForOption(_noHeaderOption),
         };
     }
 }