+++ src/QueryCat.Cli/Commands/QueryCommand.cs
@@ -40,6 +40,8 @@ internal class QueryCommand : BaseQueryCommand
             };
         var disableCacheOption = new Option<bool>("--disable-cache",
             description: "Do not use memory cache for sub-queries.");
+        var noHeaderOption = new Option<bool>("--no-header",
+            description: "Do not render header.");
 
         this.AddOption(maxErrorsOption);
         this.AddOption(statisticOption);
@@ -50,11 +52,13 @@ internal class QueryCommand : BaseQueryCommand
         this.AddOption(analyzeRowsOption);
         this.AddOption(columnsSeparatorOption);
         this.AddOption(disableCacheOption);
+        this.AddOption(noHeaderOption);
         this.SetHandler((applicationOptions, query, variables, files, queryOptions) =>
         {
             applicationOptions.InitializeLogger();
             var tableOutput = new TextTableOutput(
                 stream: StandardInputOutput.GetConsoleOutput(),
+                hasHeader: !queryOptions.NoHeader,
                 separator: queryOptions.ColumnsSeparator,
                 style: queryOptions.OutputStyle);
             var options = new ExecutionOptions
@@ -98,7 +102,8 @@ internal class QueryCommand : BaseQueryCommand
                 outputStyleOption,
                 analyzeRowsOption,
                 columnsSeparatorOption,
-                disableCacheOption)
+                disableCacheOption,
+                noHeaderOption)
             );
     }
 }